+++ orchard_fem/fruits/__init__.py
@@ -1,3 +1,0 @@
-from orchard_fem.domain import FruitAttachment
-
-__all__ = ["FruitAttachment"]